refactor: move SettingTypes to settings package
