+++ app/src/main/java/com/fpf/smartscan/settings/SettingTypes.kt
@@ -1,0 +1,11 @@
+package com.fpf.smartscan.settings
+
+object SettingTypes {
+    const val THRESHOLD = "threshold"
+    const val MODELS = "models"
+    const val MANAGE_MODELS = "manage_models"
+    const val SEARCHABLE_IMG_DIRS = "searchable_img_dirs"
+    const val SEARCHABLE_VID_DIRS = "searchable_vid_dirs"
+    const val BACKUP_RESTORE = "backup_restore"
+    const val DUPLICATES = "duplicates"
+}
