+++ app/src/main/java/com/fpf/smartscan/ui/screens/collections/CollectionsScreen.kt
@@ -50,7 +50,7 @@ import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.pluralStringResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.zIndex
-import com.fpf.smartscan.constants.Routes
+import com.fpf.smartscan.navigation.Routes
 import com.fpf.smartscan.ui.components.modals.SelectorModal
 import com.fpf.smartscan.ui.components.SlideRevealBox
 import com.fpf.smartscan.ui.components.modals.TextInputModal
